requirements.txt
astroid==2.4.2
azure-cognitiveservices-vision-computervision==0.5.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
flash==1.0.3
Flask==1.1.2
Flask-Mail==0.9.1
Flask-MySQL==1.5.1
-e git://github.com/maxcountryman/flask-uploads.git@f66d7dc93e684fa0a3a4350a38e41ae00483a796#egg=Flask_Uploads
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
isort==5.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
msrest==0.6.19
numpy==1.19.2
oauthlib==3.1.0
pdfkit==0.6.1
Pillow==7.2.0
pycparser==2.20
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
